Refactor: split play script
